Add files via upload
Image preprocessing is a crucial step in the field of image processing, aimed at improving the quality and preparing images for further analysis, such as pattern recognition, computer vision tasks, or machine learning. It involves several techniques and methods that can enhance or clean up the image to make it easier to analyze.